requirements.txt
torch>=1.8.0
pandas>=1.1.0
numpy>=1.19.0
faiss-cpu>=1.7.1
scikit-learn>=0.23.0
